Код стайл Добавлена EditingViewModel и логика работы экрана EditingScreen
