+++ FakeNFT/Scenes /Profile/ProfileScreen/ProfileViewModel.swift
@@ -6,25 +6,16 @@ protocol ProfileViewModelProtocol {
     func observeUserProfileChanges(_ handler: @escaping (UserProfile?) -> Void)
     
     func viewDidLoad()
-    func userWillcloseViewController()
-    
-    func updateName(_ name: String)
-    func updateDescription(_ description: String)
-    func updateWebSite(_ website: String)
-    
-    func photoURLdidChanged(with url: URL)
 }
 
 final class ProfileViewModel: ProfileViewModelProtocol {
     @Observable
     private(set) var userProfile: UserProfile?
     
-    private let model: ProfileService
-    private let imageValidator: ImageValidatorProtocol
+    private let profileService: ProfileService
     
-    init(model: ProfileService, imageValidator: ImageValidatorProtocol = ImageValidator()) {
-        self.model = model
-        self.imageValidator = imageValidator
+    init(model: ProfileService) {
+        self.profileService = model
         NotificationCenter.default.addObserver(self, selector: #selector(profileUpdated), name: NSNotification.Name("profileUpdated"), object: nil)
     }
     
@@ -41,82 +32,9 @@ final class ProfileViewModel: ProfileViewModelProtocol {
         fetchUserProfile()
     }
     
-    func updateName(_ name: String) {
-        guard let currentProfile = userProfile else { return }
-        userProfile = UserProfile(
-            name: name,
-            avatar: currentProfile.avatar,
-            description: currentProfile.description,
-            website: currentProfile.website,
-            nfts: currentProfile.nfts,
-            likes: currentProfile.likes,
-            id: currentProfile.id
-        )
-    }
-    
-    func updateDescription(_ description: String) {
-        guard let currentProfile = userProfile else { return }
-        userProfile = UserProfile(
-            name: currentProfile.name,
-            avatar: currentProfile.avatar,
-            description: description,
-            website: currentProfile.website,
-            nfts: currentProfile.nfts,
-            likes: currentProfile.likes,
-            id: currentProfile.id
-        )
-    }
-    
-    func updateWebSite(_ website: String) {
-        guard let currentProfile = userProfile else { return }
-        userProfile = UserProfile(
-            name: currentProfile.name,
-            avatar: currentProfile.avatar,
-            description: currentProfile.description,
-            website: website,
-            nfts: currentProfile.nfts,
-            likes: currentProfile.likes,
-            id: currentProfile.id
-        )
-    }
-    
-    func userWillcloseViewController() {
-        guard let userProfile = userProfile else { return }
-        model.updateProfile(with: userProfile) { [weak self] result in
-            guard let self = self else { return }
-            switch result {
-            case .success(let updatedProfile):
-                self.userProfile = updatedProfile
-            case .failure(let error):
-                //ToDo: - Уведомление об ошибке
-                print(error)
-            }
-        }
-    }
-    
-    func photoURLdidChanged(with url: URL) {
-        imageValidator.isValidImageURL(url) { [weak self] isValid in
-            guard let self = self else { return }
-            if isValid,
-               let currentProfile = userProfile {
-                self.userProfile = UserProfile(
-                    name: currentProfile.name,
-                    avatar: url.absoluteString,
-                    description: currentProfile.description,
-                    website: currentProfile.website,
-                    nfts: currentProfile.nfts,
-                    likes: currentProfile.likes,
-                    id: currentProfile.id
-                )
-            } else {
-                // ToDo: Уведомьте пользователя, что URL не является действительным изображением.
-            }
-        }
-    }
-    
     private func fetchUserProfile() {
         ProgressHUD.show("Загрузка...")
-        model.fetchProfile { [weak self] result in
+        profileService.fetchProfile { [weak self] result in
             guard let self = self else { return }
             switch result {
             case .success(let userProfile):
